Added middleware exception

+++ api/bs-api/bs-service/AuthorService.cs
@@ -2,6 +2,7 @@
 using bs_domain.Repositories;
 using bs_service.DTO;
 using bs_service.Mappers;
+using bs_shared.Exceptions;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -35,6 +36,10 @@ namespace bs_service
 
         public async Task<AuthorDTO> Update(AuthorDTO dto)
         {
+            var notExists = (await _repository.GetById(dto.Code.Value) is null);
+            if (notExists)
+                throw new NotFoundException($"Autor com código {dto.Code.Value} não encontrado.");
+
             var author = AuthorMapper.FromDTO(dto);
             author = await _repository.Update(author);
 
@@ -44,6 +49,9 @@ namespace bs_service
         public async Task Delete(long code)
         {
             var author = await _repository.GetById(code);
+            if (author is null)
+                throw new NotFoundException($"Autor com código {code} não encontrado.");
+
             await _repository.Delete(author);
         }
     }
