+++ api/bs-api/bs-service/SubjectService.cs
@@ -2,6 +2,7 @@
 using bs_domain.Repositories;
 using bs_service.DTO;
 using bs_service.Mappers;
+using bs_shared.Exceptions;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -35,6 +36,10 @@ namespace bs_service
 
         public async Task<SubjectDTO> Update(SubjectDTO dto)
         {
+            var notExists = (await _repository.GetById(dto.Code.Value) is null);
+            if (notExists)
+                throw new NotFoundException($"Assunto com código {dto.Code.Value} não encontrado.");
+
             var subject = SubjectMapper.FromDTO(dto);
             subject = await _repository.Update(subject);
 
@@ -44,6 +49,9 @@ namespace bs_service
         public async Task Delete(long code)
         {
             var subject = await _repository.GetById(code);
+            if (subject is null)
+                throw new NotFoundException($"Assunto com código {code} não encontrado.");
+
             await _repository.Delete(subject);
         }
     }
